Trend and Analyze Indicator Data by Attributes
The Incident Process owner at Glide Haven needs to analyze trends in resolved incidents by various attributes, such as Assignment Group, Department, Manager, or Resolution Code. 
Many Breakdowns are included with out-of-the-box Analytics and Reporting Solutions. When the desired data segmentation isn't supported out-of-the-box, a custom breakdown is required. 

Automated Breakdowns:

- Create a new custom Resolution Code Breakdown 
- Create a new Caller’s Department Breakdown 
- Run collection with Breakdown exclusions [Caller Department + Assignment Group, Category + Caller Department]
- Configure Breakdowns using the Manage Breakdowns interface and the Breakdown Matrix